new field in settings: training_type
