+++ gym_gazebo/envs/f1/modes/f1_env_manual_pilot.py
