+++ gym_gazebo/installation/formula1_laser_setup.bash
