+++ gym_gazebo/installation/formula1_setup.bash
